edit migrations, start AppslyerService, fix save subscriptions

+++ subscriptions/migrations/2019_01_25_000000_create_subscriptions_table.php
@@ -21,8 +21,8 @@ class CreateSubscriptionsTable extends Migration
             $table->string('environment');
             $table->string('original_transaction_id');
             $table->string('type');
-            $table->integer('start_date');
-            $table->integer('end_date');
+            $table->bigInteger('start_date');
+            $table->bigInteger('end_date');
             $table->text('latest_receipt');
 
         });
